bugfix/fix-slurm-cluster (#17)
* Test print the current dask config

* Set config options

* Move death timeout to cluster config

* Specify thread limit per worker

* Set process limits

* Do not nanny the process

* Do not adapt

* Remove threads hardlock value

* Add back thread limit and reduce total workers to 8

* Fix nthreads value

* Drop to a single worker

* Increase memory

* Currently working SLURM cluster

* Test high per node cpu request

* Lower core count per worker

* Drop back down to 8 cores

* Attempt with adaptive SLURM cluster

* Up maximum jobs

* Do not adapt

* Take the entire node

* Up the number of cores

* Set processes

* Test with upped processes

* Remove processes

* Move logs to CWD

* Barely increase workers

* Bump to 20 workers

* Drop to 19 workers

* Drop to 12 workers, increase death timeout to be longer than pipeline

* Drop to 10 minutes death timeout

* Get rid of death timeout all together

* Try adding chunksize parameter

* Wrong keyword, batch size instead of chunksize

* Add worker to logging

* Bump workers

* Switch to print over logging

* Drop workers to 12

* Remove processes commented out line

* Ignore dask logs

* Add directory .dask_logs to gitignore instead of file

* Add AICS pipeline data sample CSV

* Add code to skip over files that have already been generated before

* Black formatting

* Use batched map custom function rather than single map and get futures

* Change batched_map docstring

* Add error handling

* Store error message in JSON

* Fix type return from skipped cell features

* Change to adaptive SLURM cluster

* Change back to scale single time

* Do not send serialized dataset in task

* Reduce batch size to 10

* Return to working SLURM cluster settings

* Hard pin dep versions

* Pin all dep versions

* Drop column that we are merging on if it already exists to fix reruns

* Rename columns to fields in missing data exception

* Fix column names on aics data

* Import DatasetFields into download aics data script

* Use aics dask utils

* Test adaptable SLURM cluster

* Scale cluster to 60 workers

+++ actk/steps/single_cell_features/single_cell_features.py
@@ -11,9 +11,10 @@ import pandas as pd
 from aicsimageio import AICSImage
 from datastep import Step, log_run_params
 
+from aics_dask_utils import DistributedHandler
+
 from ...constants import DatasetFields
 from ...utils import dataset_utils, image_utils
-from ...utils.dask_utils import DistributedHandler
 from ..standardize_fov_array import StandardizeFOVArray
 
 ###############################################################################
@@ -35,6 +36,11 @@ class SingleCellFeaturesResult(NamedTuple):
     path: Path
 
 
+class SingleCellFeaturesError(NamedTuple):
+    cell_id: int
+    error: str
+
+
 ###############################################################################
 
 
@@ -51,36 +57,56 @@ class SingleCellFeatures(Step):
 
     @staticmethod
     def _generate_single_cell_features(
-        row_index: int, row: pd.Series, cell_ceiling_adjustment: int, save_dir: Path,
+        row_index: int,
+        row: pd.Series,
+        cell_ceiling_adjustment: int,
+        save_dir: Path,
+        overwrite: bool,
     ) -> SingleCellFeaturesResult:
-        log.info(f"Beginning Cell Feature Generation for CellId: {row.CellId}")
-
-        # Read the standardized FOV
-        image = AICSImage(row.StandardizedFOVPath)
-
-        # Select and adjust cell shape ceiling for this cell
-        adjusted = image_utils.select_and_adjust_segmentation_ceiling(
-            image=image.get_image_data("CYXZ", S=0, T=0),
-            cell_index=row.CellIndex,
-            cell_ceiling_adjustment=cell_ceiling_adjustment,
-        )
-
-        # Crop the FOV to the segmentation portions
-        cropped = image_utils.crop_raw_channels_with_segmentation(
-            image=adjusted, channels=image.get_channel_names(),
-        )
-
-        # Generate features
-        features = image_utils.get_features_from_image(cropped)
-
-        # Save to JSON
+        # Get the ultimate end save path for this cell
         save_path = save_dir / f"{row.CellId}.json"
-        with open(save_path, "w") as write_out:
-            json.dump(features, write_out)
 
-        log.info(f"Completed Cell Feature Generation for CellId: {row.CellId}")
+        # Check skip
+        if not overwrite and save_path.is_file():
+            print(f"Skipping Cell Feature Generate for Cell Id: {row.CellId}")
+            return SingleCellFeaturesResult(row.CellId, save_path)
 
-        return SingleCellFeaturesResult(row.CellId, save_path)
+        # Overwrite or didn't exist
+        print(f"Beginning Cell Feature Generation for CellId: {row.CellId}")
+
+        # Wrap errors for debugging later
+        try:
+            # Read the standardized FOV
+            image = AICSImage(row.StandardizedFOVPath)
+
+            # Select and adjust cell shape ceiling for this cell
+            adjusted = image_utils.select_and_adjust_segmentation_ceiling(
+                image=image.get_image_data("CYXZ", S=0, T=0),
+                cell_index=row.CellIndex,
+                cell_ceiling_adjustment=cell_ceiling_adjustment,
+            )
+
+            # Crop the FOV to the segmentation portions
+            cropped = image_utils.crop_raw_channels_with_segmentation(
+                image=adjusted, channels=image.get_channel_names(),
+            )
+
+            # Generate features
+            features = image_utils.get_features_from_image(cropped)
+
+            # Save to JSON
+            with open(save_path, "w") as write_out:
+                json.dump(features, write_out)
+
+            print(f"Completed Cell Feature Generation for CellId: {row.CellId}")
+            return SingleCellFeaturesResult(row.CellId, save_path)
+
+        # Catch and return error
+        except Exception as e:
+            print(
+                f"Failed Cell Feature Generation for CellId: {row.CellId}. Error: {e}"
+            )
+            return SingleCellFeaturesError(row.CellId, str(e))
 
     @log_run_params
     def run(
@@ -88,7 +114,7 @@ class SingleCellFeatures(Step):
         dataset: Union[str, Path, pd.DataFrame, dd.DataFrame],
         cell_ceiling_adjustment: int = 7,
         distributed_executor_address: Optional[str] = None,
-        clean: bool = False,
+        overwrite: bool = False,
         debug: bool = False,
         **kwargs,
     ):
@@ -112,9 +138,10 @@ class SingleCellFeatures(Step):
             An optional executor address to pass to some computation engine.
             Default: None
 
-        clean: bool
-            Should the local staging directory be cleaned prior to this run.
-            Default: False (Do not clean)
+        overwrite: bool
+            If this step has already partially or completely run, should it overwrite
+            the previous files or not.
+            Default: False (Do not overwrite or regenerate files)
 
         debug: bool
             A debug flag for the developer to use to manipulate how much data runs,
@@ -130,13 +157,13 @@ class SingleCellFeatures(Step):
         if isinstance(dataset, (str, Path)):
             dataset = Path(dataset).expanduser().resolve(strict=True)
 
-            # Read dataset
-            dataset = pd.read_csv(dataset)
+            # Check dataset and manifest have required fields
+            dataset_utils.check_required_fields(
+                dataset=dataset, required_fields=REQUIRED_DATASET_FIELDS,
+            )
 
-        # Check dataset and manifest have required fields
-        dataset_utils.check_required_fields(
-            dataset=dataset, required_fields=REQUIRED_DATASET_FIELDS,
-        )
+        # Read dataset
+        dataset = pd.read_csv(dataset)
 
         # Create features directory
         features_dir = self.step_local_staging_dir / "cell_features"
@@ -145,7 +172,7 @@ class SingleCellFeatures(Step):
         # Process each row
         with DistributedHandler(distributed_executor_address) as handler:
             # Start processing
-            futures = handler.client.map(
+            results = handler.batched_map(
                 self._generate_single_cell_features,
                 # Convert dataframe iterrows into two lists of items to iterate over
                 # One list will be row index
@@ -155,23 +182,31 @@ class SingleCellFeatures(Step):
                 # mapped function call
                 [cell_ceiling_adjustment for i in range(len(dataset))],
                 [features_dir for i in range(len(dataset))],
+                [overwrite for i in range(len(dataset))],
             )
-
-            # Block until all complete
-            results = handler.gather(futures)
 
         # Generate features paths rows
         cell_features_dataset = []
+        errors = []
         for result in results:
-            cell_features_dataset.append(
-                {
-                    DatasetFields.CellId: result.cell_id,
-                    DatasetFields.CellFeaturesPath: result.path,
-                }
-            )
+            if isinstance(result, SingleCellFeaturesResult):
+                cell_features_dataset.append(
+                    {
+                        DatasetFields.CellId: result.cell_id,
+                        DatasetFields.CellFeaturesPath: result.path,
+                    }
+                )
+            else:
+                errors.append(
+                    {DatasetFields.CellId: result.cell_id, "Error": result.error}
+                )
 
         # Convert features paths rows to dataframe
         cell_features_dataset = pd.DataFrame(cell_features_dataset)
+
+        # Drop CellFeaturesPath column if it already exists
+        if DatasetFields.CellFeaturesPath in dataset.columns:
+            dataset = dataset.drop(columns=[DatasetFields.CellFeaturesPath])
 
         # Join original dataset to the fov paths
         self.manifest = dataset.merge(cell_features_dataset, on=DatasetFields.CellId)
@@ -179,5 +214,9 @@ class SingleCellFeatures(Step):
         # Save manifest to CSV
         manifest_save_path = self.step_local_staging_dir / f"manifest.csv"
         self.manifest.to_csv(manifest_save_path, index=False)
+
+        # Save errored cells to JSON
+        with open(self.step_local_staging_dir / "errors.json", "w") as write_out:
+            json.dump(errors, write_out)
 
         return manifest_save_path
